zajel_backend/requirements.txt
Django==5.2.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8

# PostgreSQL (optional - uncomment for production)
# psycopg2-binary==2.9.9 